replaced useMemo to useRef.
Bcos of this:
React.memo only checks for prop changes. If your function component wrapped in React.memo has a useState or useContext Hook in its implementation, it will still rerender when state or context change.

And this:
A ref created with useRef will be created only when the component has been mounted and preserved for the full lifecycle.